Optimised subset of LHC Sampling Plan

Optimised subset of LHC Sampling Plan

Generate an optimised subset of an existing plan. The optimisation of the subset is based on a genetic algorithm.

Functions

LatinHypercubeSampling.subLHCoptim — Method.

function subLHCoptim(X,n::Int,gens;popsize::Int=100,ntour::Int=2,ptour=0.8)

Produce an optimized Latin Hyper Cube with n sample points from a subset of points in X. Optimization is run for gens generations.

source

LatinHypercubeSampling.subLHCindex — Method.

function subLHCindex(X,Xsub)

Index in the large LHC to get the subLHC.

source

Example

Create an optimised subset LHC plan from an existing plan with 120 points in 2 dimensions. alt text

julia> subLHCoptim(X,Xsub)

alt text

The indices of the subset in the larger plan can be extracted with

julia> subLHCindex(X,Xsub)